Fix several bugs with replace and replace around steps (#3)
* Expand replace around slice so that replacements are always flat

* Fix several bugs with replace and replace around steps

Properly handles list sink commands by treating them like
a replace of the current and previous list item with a
new list item containing the sinked list.

Properly handles arbitrary-depth textblocks with adjacent
marks with the same id. This means that deletions across
list items can now be applied correctly.

"Skips" deletes on zero-width spaces, to avoid having to
double-backspace at the end of a textblock with a zero-width
space.